% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveFace.R
\name{aveFace}
\alias{aveFace}
\title{Average Face of Training Data Set}
\usage{
aveFace(x, display = TRUE)
}
\arguments{
\item{x}{a data frame consisting the suitability scores of a given characteristics
(terrain, soil, water and temperature) for a 
given crop (e.g. coconut, cassava, etc.);}

\item{method}{the method for computing the overall suitability, which includes the
\code{"minimum"}, \code{"maximum"}, \code{"sum"}, \code{"product"},
\code{"average"}, \code{"exponential"} and \code{"gamma"}. If \code{NULL},
\code{"minimum"} is used.}

\item{interval}{if \code{NULL}, the interval used are the following: 0-25% (Not
suitable, N), 25%-50% (Marginally Suitable, S3), 50%-75% (Moderately Suitable, S2), and
75%-100% (Highly Suitable, S1). But users can assign a custom intervals by specificying
the values of the end points of the intervals.}

\item{output}{the output to be returned, either the scores or class. If \code{NULL},
both are returned.}
}
\description{
This function uses PCA to compute the eigenfaces and recognizes the data.
}
\examples{
To reproduce this example, please refer to \\code{https://github.com/alstat/rface}

# Call the Face Recognition package
library(rface)

# Import all images in the directory "data/jaffe/"
imgData <- importImages("~/Downloads/jaffe_training/", display = FALSE)

# Take the average of the faces
img_mean <- aveFace(imgData, display = FALSE)
showFace(img_mean)
}

