% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brStick.R
\name{brStick}
\alias{brStick}
\title{Broken Stick Stopping Rule}
\usage{
brStick(x, plot = FALSE)
}
\arguments{
\item{x}{numeric, a vector of the eigenvalues of the principal components.}

\item{plot}{boolean default is FALSE, toggle plot.}
}
\description{
This function is a criterion for choosing the number of principal components.
}
\examples{
summary(pc.cr <- princomp(USArrests, cor = TRUE))
brStick(pc.cr$sdev ** 2)

# Show plot
brStick(pc.cr$sdev ** 2, plot = TRUE)
}
\seealso{
\code{\link{rbrStick}}
}

