% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importImages.R
\name{importImages}
\alias{importImages}
\title{Import Images}
\usage{
importImages(path = ".", display = FALSE)
}
\arguments{
\item{path}{string input; directory of the folder, default is set to the working directory.}

\item{display}{boolean input; toggle display of the images.}
}
\description{
Import images from the folder.
}
\details{
The \code{path} of the folder should end with \code{/}, so that function will import all
filenames following that symbol.
}
\examples{
library(fr)

# Import all images in the directory "data/jaffe/"
imgData <- importImages("~/Documents/Data/jaffe/")
showFace(imgData) # display the images
}
\seealso{
\code{\link{EBImage::readImage}}
}

