% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbrStick.R
\name{rbrStick}
\alias{rbrStick}
\title{Relative Broken Stick Stopping Rule}
\usage{
rbrStick(x, plot = FALSE)
}
\arguments{
\item{x}{numeric, a vector of the eigenvalues of the principal components.}

\item{plot}{boolean default is FALSE, toggle plot.}
}
\description{
This function is a criterion for choosing the number of principal components.
}
\examples{
summary(pc.cr <- princomp(USArrests, cor = TRUE))
rbrStick(pc.cr$sdev ** 2)

# Show plot
rbrStick(pc.cr$sdev ** 2, plot = TRUE)
}
\seealso{
\code{\link{brStick}}
}

