% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recognize.R
\name{recognize}
\alias{recognize}
\title{Recognize Image (Human Face)}
\usage{
recognize(x, y, rule = "simple f-share", display = TRUE)
}
\arguments{
\item{x}{string input; image file input to be recognize.}

\item{y}{model supervised on learning from the training dataset.}

\item{rule}{string input; stopping rule for choosing number of principal components.
Options are: \code{'simple f-share'} for simple fair-share; \code{'broken stick'}; and,
\code{'rel-broken stick'} for relative broken stick stopping rule.}

\item{display}{displays the input image, and the corresponding output image --
as recognized by the model.}
}
\description{
Function for recognizing the image using the model trained from the data.
}
\examples{
To reproduce this example, please refer to \\code{https://github.com/alstat/rface}

# Call the Face Recognition package
library(rface)
# Import all images in the directory "data/jaffe/"
imgData <- importImages("~/Downloads/jaffe_training/", display = FALSE)

# Show imported images
showFace(imgData)

# Take the average of the faces
img_mean <- aveFace(imgData, display = FALSE)

# Show the average of the faces
showFace(img_mean)

# Train the algorithm for the images
model <- learn(imgData)

# Show the eigen faces
showFace(model)

# Input an image and recognize it
recognize("~/Downloads/jaffe_testing/KA.AN1.39.tiff", model, display = TRUE, rule = "simple f-share")
}

