% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn}
\alias{learn}
\title{Supervised Learning}
\usage{
learn(x)
}
\arguments{
\item{x}{matrix input of class \code{images}; this object contains the imported images
transformed into matrix by \code{\link{importImages}} function.}
}
\description{
Supervised the algorithm to learn from the training data set.
}
\examples{
To reproduce this example, please refer to \\code{https://github.com/alstat/rface}

# Call the Face Recognition package
library(rface)
# Import all images in the directory "data/jaffe/"
imgData <- importImages("~/Downloads/jaffe_training/", display = FALSE)

# Train the algorithm for the images
model <- learn(imgData)

# Show the eigen faces
showFace(model)
}

