% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recognize.R
\name{recognize}
\alias{recognize}
\title{Recognize Image (Human Face)}
\usage{
recognize(x, y, rule = "simple f-share", display = TRUE)
}
\arguments{
\item{x}{string input; image file input to be recognize.}

\item{y}{model supervised on learning from the training dataset.}

\item{rule}{string input; stopping rule for choosing number of principal components.
Options are: \code{'simple f-share'} for simple fair-share; \code{'broken stick'}; and,
\code{'rel-broken stick'} for relative broken stick stopping rule.}

\item{display}{displays the input image, and the corresponding output image --
as recognized by the model.}
}
\description{
Function for recognizing the image using the model trained from the data.
}
\examples{
library(fr)

# Import all images in the directory "data/jaffe/"
imgData <- importImges("~/Documents/Data/jaffe/", display = FALSE)

model <- learn(imgData)
a <- recognize("~/Documents/Data/jaffe/KA.FE3.47.tiff", criterion, display = TRUE, rule = "rel-broken stick")
}

