% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rface-package.R
\docType{package}
\name{rface-package}
\alias{rface-package}
\title{Face Recognition using Principal Component Analysis Package}
\description{
R package for face and pattern recognition using techniques such as
principal componanent analysis and neural networks.
}
\author{
Al-Ahmadgaid B. Asaad
}

